Revert "Removed need for `primitives.meta`."
This reverts commit 718a1c30fa45c16b61ea30e23ebc211a09917a84.

+++ electrical/system/defs.h
@@ -23,9 +23,9 @@
 
 struct CMSISPutTuple
 {
-    volatile u32* dst;
-    i32           pos;
-    u32           msk;
+    volatile long unsigned int* dst;
+    i32                         pos;
+    u32                         msk;
 };
 
 static mustinline void
